#6 feat: 3단계 눈사람 애니메이션 HomeFragment로부터 분리
