+++ app/src/main/java/com/ariari/mowoori/ui/home/HomeViewModel.kt
@@ -39,9 +39,6 @@ class HomeViewModel @Inject constructor(
     private var _snowmanLevel = MutableLiveData<SnowmanLevel>()
     val snowmanLevel: LiveData<SnowmanLevel> = _snowmanLevel
 
-    private var _isFirstCycle = true
-    val isFirstCycle = _isFirstCycle
-
     private val snowAnimList: MutableList<Animator> = mutableListOf()
 
     fun setUserInfo() {
@@ -85,10 +82,6 @@ class HomeViewModel @Inject constructor(
         homeRepository.setCurrentGroupId(groupId)
     }
 
-    fun setIsFirstCycle(isFirst: Boolean) {
-        _isFirstCycle = isFirst
-    }
-
     fun updateIsSnowing() {
         if (isSnowing.value == null) {
             _isSnowing.postValue(true)
@@ -104,6 +97,12 @@ class HomeViewModel @Inject constructor(
     fun addSnowAnim(anim: Animator) {
         if (!snowAnimList.contains(anim)) {
             snowAnimList.add(anim)
+        }
+    }
+
+    fun addSnowAnim(animators: List<Animator>) {
+        animators.forEach { anim ->
+            addSnowAnim(anim)
         }
     }
