+++ app/src/main/java/com/ariari/mowoori/ui/home/animator/WinterAnimatorLv1.kt
@@ -1,0 +1,4 @@
+package com.ariari.mowoori.ui.home.animator
+
+class WinterAnimatorLv1 {
+}
